requirements.txt
uv
streamlit
openai
pandas
matplotlib
python-dotenv
langchain
langchain-openai
langchain-experimental
langchain-community
openpyxl
tabulate
plotly